+++ webui/webapp/Dashboard/src/org/netxms/ui/eclipse/dashboard/widgets/internal/AbstractChartConfig.java
@@ -56,10 +56,10 @@ public abstract class AbstractChartConfig extends DashboardElementConfig
 	private int refreshRate = 30;
    
    @Element(required = false)
-   private int minYScaleValue = 0;
+   private double minYScaleValue = 0;
 
    @Element(required = false)
-   private int maxYScaleValue = 100;
+   private double maxYScaleValue = 100;
 
    @Element(required = false)
    private boolean modifyYBase = false;
@@ -221,7 +221,7 @@ public abstract class AbstractChartConfig extends DashboardElementConfig
    /**
     * @return the minYScaleValue
     */
-   public int getMinYScaleValue()
+   public double getMinYScaleValue()
    {
       return minYScaleValue;
    }
@@ -229,7 +229,7 @@ public abstract class AbstractChartConfig extends DashboardElementConfig
    /**
     * @param minYScaleValue the minYScaleValue to set
     */
-   public void setMinYScaleValue(int minYScaleValue)
+   public void setMinYScaleValue(double minYScaleValue)
    {
       this.minYScaleValue = minYScaleValue;
    }
@@ -237,7 +237,7 @@ public abstract class AbstractChartConfig extends DashboardElementConfig
    /**
     * @return the maxYScaleValue
     */
-   public int getMaxYScaleValue()
+   public double getMaxYScaleValue()
    {
       return maxYScaleValue;
    }
@@ -245,16 +245,22 @@ public abstract class AbstractChartConfig extends DashboardElementConfig
    /**
     * @param maxYScaleValue the maxYScaleValue to set
     */
-   public void setMaxYScaleValue(int maxYScaleValue)
+   public void setMaxYScaleValue(double maxYScaleValue)
    {
       this.maxYScaleValue = maxYScaleValue;
    }
    
+   /**
+    * @return
+    */
    public boolean isAutoScale()
    {
       return autoScale;
    }
 
+   /**
+    * @param autoScale
+    */
    public void setAutoScale(boolean autoScale)
    {
       this.autoScale = autoScale;
